Add an overlay loader loader for disk images to work with a bootloader.
If a program needs overlays but doesn't want to require a specific ROM,
just that the bootloader is compatible with the standard, then it must
load its own overlay loader as the first thing it does.

This commit adds a code snippet that does that, and then loads the
first thing after it in the disk image, on the assumption that it is an
overlay that starts the main program.

Thus, this snippet should be suitable as a paste-in prefix to get an
overlay loader even if it wasn't in ROM.

Note that this snippet requires the preprocessor script.