Added an attempt at an overlay based version of the assembler.
This is multipass, checking and transforming in several steps so as to
fit each task into one overlay, for speed and code simplicity.

However, I've gotten as far as finding and removing comments, and the
code for that is larger than one overlay and I'm having trouble making
it smaller. Plus the code is already fairly large without even finding
any instructions yet, nevermind parsing arguments.

So I'm not sure if this is any more feasible than the other version.

(Plus I just realized NILLIST means the object code may be larger than
the source code, so the overwriting I'm doing may cause problems too.)