Fixed some problems in the overlay loader and its loader.
Previously, the overlay loader only supported RAM sizes up to 256.
With this fix it supports the entire 24-bit address range.
(The problem was that it didn't reset the entire address to 0.)

The loader loader had an extra unnecessary instruction, now removed.
In addition, its overlay loader looping address now uses a label with
@relative for clarity of intent.