Added an assembler written in Bash script.
This assembler is somewhat more relaxed than Senbir's when it comes to
formatting, but expects the same arguments as it, written the same way.

It is also fairly flexible when it comes to output format, supporting
not only what I believe Senbir's disk image format to be, but also hex,
JSON, a pretty-printed JS array format, and _some_ custom formats.

Instruction-wise, it supports those that are in my version of Senbir,
but not (yet) those added later; though they should not be hard to add.