Added a TC-06 emulator written in TC-06 assembly.
While not comprehensively tested, it appears to work (though slowly).

As-is it doesn't really give much, but with some (relatively minor)
changes it could be used to do things like emulate a machine with more
memory than the host machine has, or even do preemptive multitasking
(if given enough disk space for two programs, at least).