Added a (partial) VHDL implementation of the RISC-06 architecture.
Most of the components are there, and most of the instructions are
implemented (just missing SPC, mul, div, mod and exp).

However, the "monitor" device does not actually output anything to any
real display, the "disk drive" is not persistent (it's just another
block of RAM) and the RAM is not being initialized from the boot ROM.

I also haven't actually tested it. (My VHDL toolchain is not working
properly right now; though it compiles, it fails to simulate anything.)